src/main/java/dp/Solution300.java
package dp;

/**
 * @author qpzm7903
 * @since 2020-03-14-8:09
 */

public class Solution300 {
    public int lengthOfLIS(int[] nums) {
        if (nums.length == 0) {
            return 0;
        }
        int[] dp = new int[nums.length];
        dp[0] = 1;
        int maxAns = 1;
        for (int i = 1; i < nums.length; i++) {
            int maxLength = 0;
            for (int j = 0; j < i; j++) {
                if (nums[i] > nums[j]) {
                    maxLength = Math.max(maxLength, dp[j]);
                }
            }
            dp[i] = maxLength + 1;
            maxAns = Math.max(dp[i], maxAns);
        }
        return maxAns;
    }
}
